file relocation
