+++ src/main/java/com/teamchallenge/easybuy/dto/user/CustomerProfileDto.java
@@ -1,4 +1,4 @@
-package com.teamchallenge.easybuy.dto;
+package com.teamchallenge.easybuy.dto.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
