+++ src/main/java/com/teamchallenge/easybuy/dto/user/CustomerProfileResponseDto.java
@@ -1,5 +1,7 @@
-package com.teamchallenge.easybuy.dto;
+package com.teamchallenge.easybuy.dto.user;
 
+import com.teamchallenge.easybuy.dto.AddressDto;
+import com.teamchallenge.easybuy.dto.user.CustomerProfileDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
