+++ src/main/java/com/teamchallenge/easybuy/repo/user/CustomerRepository.java
@@ -1,6 +1,6 @@
-package com.teamchallenge.easybuy.repo;
+package com.teamchallenge.easybuy.repo.user;
 
-import com.teamchallenge.easybuy.models.Customer;
+import com.teamchallenge.easybuy.models.user.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
